Add files via upload
Change the time step by changing the time_step parameter value, change the solver type by flipping the switch, error is output on the scope rn but still need to put in something to find the max error, also something for computation time. Feel free to modify or add to this; tbh the description made it sound like we were supposed to just do this in Matlab with no Simulink but based on what we did on Friday I went ahead with this.